INFUND-3760 change double to Decimal(5,2) and various other PR fixes

+++ ifs-data-service/src/main/resources/db/migration/V42_0__Add_completion_to competition_and_application.sql
@@ -1,2 +1,2 @@
 ALTER TABLE `application`
-ADD COLUMN `completion` DOUBLE NOT NULL DEFAULT 0 AFTER `assessor_feedback_file_entry_id`;
+ADD COLUMN `completion` DECIMAL(5,2) NOT NULL DEFAULT 0 AFTER `assessor_feedback_file_entry_id`;
